Add files via upload
The code demonstrates SHA-256 hashing and AES-GCM encryption/decryption, brute-force function highlights the importance of string passwords and large keyspaces.
It also provides Benchmarking performance by providing timing information for cryptographic operations which are useful understanding their computational costs
The above code performs several cryptographic operations, find below the detailed breakdown of each function.

--------------------------------

1) "findPassword"

This function attempts a brute-force password attack to decrypt a ciphertext using the AES-GCM encryption algorithm with a SHA-256 derived key.

It take an input of 12-byte nonce which are all initialized to zeros, a hardcoded data string and a hardcoded ciphertext and MAC 

For Brute-forcing, it iterates over all possible 4 digit numeric passwords i.e., 0000 to 9999.
For each password, it will compute a SHA-256 hash of the password to generate a 256-bit key, it will create a AES cipher block using the key and an AES-GCM instance with the cipher block. It will try to decrypt the ciphermac with the nonce, ciphertext and associated data.  If the decryption succeeds, it prints the correct password and the plaintext and breaksout of the loop

-------------------------------
2) "timedeko"

This function measures the time it takes to compute the SHA-256 hash of the given message by taking input a hardcoded string and initializing the timer. 
It will then compute the SHA-256 hash of the input message and stops the timer and calculate the duration. It will then print the hash and the time taken in nanoseconds.
The main purpose of adding this function is for the benchmark of hashing process.

----------------------------------

3) "onembtime"

This function measures the time taken to encrypt and decrypt a 1MB random message using AES_GCM.
It will generate a random 1MB message and a random 256-bit key, and will create an AES-GCM instance with a randomly generated nonce.
The function will then encrypt the 1MB message using 'aesgcm.Seal' and will measure the time taken and then print the ciphertext size.
To decrypt the ciphertext, it will use 'aesgcm.Open' and measure the time taken, and then will print the plaintext size.

This function provides performance metrics for AES-GCM encryption and decryption for a large data size.